moved window and taper to algorithm module
