+++ cxx/python/seismic/seismic_py.cc
@@ -9,12 +9,10 @@
 
 #include <mspass/seismic/keywords.h>
 #include <mspass/seismic/SlownessVector.h>
-#include <mspass/seismic/TimeWindow.h>
 #include <mspass/seismic/TimeSeries.h>
 #include <mspass/seismic/Seismogram.h>
 #include <mspass/seismic/Ensemble.h>
 #include <mspass/seismic/PowerSpectrum.h>
-#include <mspass/seismic/Taper.h>
 
 #include "python/utility/Publicdmatrix_py.h"
 #include "python/utility/boost_any_converter_py.h"
@@ -95,28 +93,6 @@ public:
   }
 };
 
-/* Trampoline class for BasicTaper - used for family of taper classes*/
-class PyBasicTaper : public BasicTaper
-{
-public:
-  int apply(TimeSeries& d)
-  {
-    PYBIND11_OVERLOAD_PURE(
-      int,
-      BasicTaper,
-      TimeSeries&,
-      d);
-  }
-  int apply(Seismogram& d)
-  {
-    PYBIND11_OVERLOAD_PURE(
-      int,
-      BasicTaper,
-      Seismogram&,
-      d);
-  }
-};
-
 PYBIND11_MODULE(seismic, m) {
   m.attr("__name__") = "mspasspy.ccore.seismic";
   m.doc() = "A submodule for seismic namespace of ccore";
@@ -189,16 +165,6 @@ PYBIND11_MODULE(seismic, m) {
     .def("baz",&SlownessVector::baz,"Return the so called back azimuth defined by a slowness vector")
     .def_readwrite("ux",&SlownessVector::ux,"Slowness component in the x (Easting) direction")
     .def_readwrite("uy",&SlownessVector::uy,"Slowness component in the y (Northing) direction")
-  ;
-
-  py::class_<TimeWindow>(m,"TimeWindow","Simple description of a time window")
-    .def(py::init<>(),"Default constructor")
-    .def(py::init<const double, const double>(),"Construct from start and end time")
-    .def(py::init<const TimeWindow&>(),"Copy constuctor")
-    .def("shift",&TimeWindow::shift,"Shift the reference time by a specified number of seconds")
-    .def("length",&TimeWindow::length,"Return the size of the window in seconds")
-    .def_readwrite("start",&TimeWindow::start,"Start time of the window")
-    .def_readwrite("end",&TimeWindow::end,"End time of the window")
   ;
 
   py::enum_<TimeReferenceType>(m,"TimeReferenceType")
@@ -655,49 +621,6 @@ PYBIND11_MODULE(seismic, m) {
       .def_readonly("spectrum",&PowerSpectrum::spectrum,
           "Vector containing estimated power spectrum; equally spaced ordered in increasing frequency")
       .def_readwrite("elog",&PowerSpectrum::elog,"Handle to ErrorLogger")
-    ;
-    py::class_<BasicTaper,PyBasicTaper>(m,"BasicTaper",
-                     "Base class for family of taper algorithms")
-      .def(py::init<>())
-      /* note sure virtual classes need a definition here */
-      /*
-      .def("apply",py::overload_cast<TimeSeries&>(&BasicTaper::apply),
-         "TimeSeries overload of base class")
-      .def("apply",py::overload_cast<Seismogram&>(&BasicTaper::apply))
-      */
-    ;
-    py::class_<LinearTaper,BasicTaper>(m,"LinearTaper",
-       "Define a ramp taper function")
-      .def(py::init<>())
-      .def(py::init<const double, const double, const double, const double>())
-      .def("apply",py::overload_cast<TimeSeries&>(&LinearTaper::apply),"Apply taper to a scalar TimeSeries object")
-      .def("apply",py::overload_cast<Seismogram&>(&LinearTaper::apply),"Apply taper to a Seismogram (3C) object")
-    ;
-    py::class_<CosineTaper,BasicTaper>(m,"CosineTaper",
-       "Define a taper using a half period cosine function")
-      .def(py::init<>())
-      .def(py::init<const double, const double, const double, const double>())
-      .def("apply",py::overload_cast<TimeSeries&>(&CosineTaper::apply),"Apply taper to a scalar TimeSeries object")
-      .def("apply",py::overload_cast<Seismogram&>(&CosineTaper::apply),"Apply taper to a Seismogram (3C) object")
-    ;
-    py::class_<VectorTaper,BasicTaper>(m,"VectorTaper",
-       "Define generic taper function with a parallel vector of weights")
-      .def(py::init<>())
-      .def(py::init<const std::vector<double>>())
-      .def("apply",py::overload_cast<TimeSeries&>(&VectorTaper::apply),
-        "Apply taper to a scalar TimeSeries object")
-      .def("apply",py::overload_cast<Seismogram&>(&VectorTaper::apply),
-         "Apply taper to a Seismogram (3C) object")
-    ;
-    py::class_<TopMute>(m,"_TopMute",
-      "Defines a top mute operator (one sided taper")
-      .def(py::init<>())
-      .def(py::init<const double, const double, const std::string>())
-      .def(py::init<const TopMute&>())
-      .def("apply",py::overload_cast<TimeSeries&>(&TopMute::apply),
-        "Apply to a TimeSeries object")
-      .def("apply",py::overload_cast<Seismogram&>(&TopMute::apply),
-        "Apply to a Seismogram object")
     ;
 }
