+++ cxx/src/lib/algorithms/tseries_helpers.cc
@@ -1,6 +1,6 @@
 #include <math.h>
 #include "mspass/utility/MsPASSError.h"
-#include "mspass/seismic/TimeWindow.h"
+#include "mspass/algorithms/TimeWindow.h"
 #include "mspass/seismic/TimeSeries.h"
 #include "mspass/seismic/Ensemble.h"
 namespace mspass::algorithms
